feat: implement migration handling with new migrator model
