+++ infra/controller.js
@@ -1,19 +1,14 @@
-import { join } from "node:path";
-import migrationsRunner from "node-pg-migrate";
-import database from "infra/database.js";
 import { InternalServerError, MethodNotAllowedError } from "infra/errors.js";
 
-const defaultMigrationsRunner = {
-  dir: join(process.cwd(), "infra", "migrations"),
-  direction: "up",
-  verbose: true,
-  migrationsTable: "pgmigrations",
-};
-
 function onErrorHandler(error, request, response) {
+  if (error.name === "ServiceError") {
+    return response.status(error.statusCode).json(error);
+  }
+
   const publicErrorObject = new InternalServerError({
     statusCode: error.statusCode,
     cause: error,
+    message: error.message,
   });
 
   response.status(publicErrorObject.statusCode).json(publicErrorObject);
@@ -24,49 +19,11 @@ function onNoMatchHandler(request, response) {
   response.status(publicErrorObject.statusCode).json(publicErrorObject);
 }
 
-async function runPendingMigrations() {
-  let dbClient;
-
-  try {
-    dbClient = await database.getNewClient();
-
-    const migratedMigrations = await migrationsRunner({
-      ...defaultMigrationsRunner,
-      dbClient: dbClient,
-      dryRun: false,
-    });
-
-    return migratedMigrations;
-  } finally {
-    await dbClient?.end();
-  }
-}
-
-async function listPendingMigrations() {
-  let dbClient;
-
-  try {
-    dbClient = await database.getNewClient();
-
-    const pendingMigrations = await migrationsRunner({
-      ...defaultMigrationsRunner,
-      dbClient: dbClient,
-      dryRun: true,
-    });
-
-    return pendingMigrations;
-  } finally {
-    dbClient?.end();
-  }
-}
-
 const controller = {
   errorHandlers: {
     onError: onErrorHandler,
     onNoMatch: onNoMatchHandler,
   },
-  listPendingMigrations,
-  runPendingMigrations,
 };
 
 export default controller;
